pyproject.toml
[project]
name = "takenoko-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
